Organize tests & File/Folder Structure
